Renamed track controls, fixed soloing
